refactor: player 사용 로직 일부 이동
